yolov6_requirements.txt
absl-py==1.4.0
addict==2.4.0
cachetools==5.3.0
# certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==3.1.0
cmake==3.26.1
coloredlogs==15.0.1
contourpy==1.0.7
cycler==0.11.0
filelock==3.10.7
flatbuffers==23.3.3
fonttools==4.39.3
google-auth==2.17.1
google-auth-oauthlib==0.4.6
grpcio==1.53.0
humanfriendly==10.0
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
Jinja2==3.1.2
kiwisolver==1.4.4
lit==16.0.0
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
mpmath==1.3.0
networkx==3.0
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
onnx==1.13.1
onnx-simplifier==0.4.19
onnxruntime==1.14.1
opencv-python==4.7.0.72
packaging==23.0
Pillow==9.4.0
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.3.3
rsa==4.9
scipy==1.10.1
six==1.16.0
sympy==1.11.1
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
thop==0.1.1.post2209072238
torch==2.0.0
torchsummary==1.5.1
torchvision==0.15.1
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
zipp==3.15.0
